refactor: extract logging to RequestLogger class
Move all logging responsibility (9 methods, 7
properties) from UpdateServer to dedicated
Logging\RequestLogger class. Fixes ClassTooLong
and CognitiveComplexity violations. UpdateServer
drops from ~960 to 782 lines.
