+++ src/UpdateServer.php
@@ -9,6 +9,7 @@ use Apermo\WpUpdateServer\Auth\LicenseProvider;
 use Apermo\WpUpdateServer\Cache\CacheInterface;
 use Apermo\WpUpdateServer\Cache\FileCache;
 use Apermo\WpUpdateServer\Exception\InvalidPackageException;
+use Apermo\WpUpdateServer\Logging\RequestLogger;
 use RuntimeException;
 
 /**
@@ -19,16 +20,10 @@ use RuntimeException;
  */
 class UpdateServer {
 
-	/** Date format for daily log rotation. */
-	public const FILE_PER_DAY = 'Y-m-d';
-
-	/** Date format for monthly log rotation. */
-	public const FILE_PER_MONTH = 'Y-m';
-
-	/** Allowed release channel identifiers. */
+	/** @var string[] Allowed release channel identifiers. */
 	private const VALID_CHANNELS = [ 'stable', 'rc', 'beta', 'alpha' ];
 
-	/** Actions that do not require a package slug. */
+	/** @var string[] Actions that do not require a package slug. */
 	private const SLUG_OPTIONAL_ACTIONS = [ 'composer_packages', 'upload' ];
 
 	/**
@@ -60,34 +55,6 @@ class UpdateServer {
 	protected array $assetDirectories = [];
 
 	/**
-	 * Absolute path to the log directory.
-	 *
-	 * @var string
-	 */
-	protected string $logDirectory;
-
-	/**
-	 * Whether log file rotation is enabled.
-	 *
-	 * @var bool
-	 */
-	protected bool $logRotationEnabled = false;
-
-	/**
-	 * Date format suffix appended to rotated log filenames.
-	 *
-	 * @var ?string
-	 */
-	protected ?string $logDateSuffix = null;
-
-	/**
-	 * Maximum number of rotated log files to keep.
-	 *
-	 * @var int
-	 */
-	protected int $logBackupCount = 0;
-
-	/**
 	 * Cache backend for package metadata.
 	 *
 	 * @var CacheInterface
@@ -107,6 +74,13 @@ class UpdateServer {
 	 * @var PackageRepository
 	 */
 	protected PackageRepository $packageRepository;
+
+	/**
+	 * Request logger instance.
+	 *
+	 * @var RequestLogger
+	 */
+	protected RequestLogger $logger;
 
 	/**
 	 * Public base URL of the update server.
@@ -130,27 +104,6 @@ class UpdateServer {
 	protected mixed $packageFileLoader = [ Package::class, 'fromArchive' ];
 
 	/**
-	 * Whether IP addresses are anonymized before logging.
-	 *
-	 * @var bool
-	 */
-	protected bool $ipAnonymizationEnabled = false;
-
-	/**
-	 * Binary mask for anonymizing IPv4 addresses.
-	 *
-	 * @var string
-	 */
-	protected string $ip4Mask = '';
-
-	/**
-	 * Binary mask for anonymizing IPv6 addresses.
-	 *
-	 * @var string
-	 */
-	protected string $ip6Mask = '';
-
-	/**
 	 * Create a new instance.
 	 *
 	 * @param string|null $serverUrl      Public base URL; auto-detected when null.
@@ -168,7 +121,6 @@ class UpdateServer {
 
 		$this->serverUrl = $serverUrl;
 		$this->packageDirectory = $serverDirectory . '/packages';
-		$this->logDirectory = $serverDirectory . '/logs';
 
 		$this->bannerDirectory = $serverDirectory . '/package-assets/banners';
 		$this->assetDirectories = [
@@ -176,10 +128,8 @@ class UpdateServer {
 			'icons'   => $serverDirectory . '/package-assets/icons',
 		];
 
-		$this->ip4Mask = \pack( 'H*', 'ffffff00' );
-		$this->ip6Mask = \pack( 'H*', 'ffffffffffff00000000000000000000' );
-
 		$this->cache = new FileCache( $serverDirectory . '/cache' );
+		$this->logger = new RequestLogger( $serverDirectory . '/logs' );
 		$this->config = Config::fromFile( $this->serverDirectory . '/config.php' );
 		$this->packageRepository = new PackageRepository(
 			$this->packageDirectory,
@@ -262,12 +212,12 @@ class UpdateServer {
 	 */
 	protected function applyConfig(): void {
 		if ( $this->config->get( 'logging.anonymize_ip', false ) ) {
-			$this->enableIpAnonymization();
+			$this->logger->enableIpAnonymization();
 		}
 
 		if ( $this->config->get( 'logging.rotation.enabled', false ) ) {
-			$this->enableLogRotation(
-				$this->config->get( 'logging.rotation.period', self::FILE_PER_MONTH ),
+			$this->logger->enableLogRotation(
+				$this->config->get( 'logging.rotation.period', RequestLogger::FILE_PER_MONTH ),
 				(int) $this->config->get( 'logging.rotation.keep', 10 ),
 			);
 		}
@@ -293,7 +243,7 @@ class UpdateServer {
 		$this->startTime = \microtime( true );
 
 		$request = $this->initRequest( $query, $headers );
-		$this->logRequest( $request );
+		$this->logger->log( $request );
 
 		$this->loadPackageFor( $request );
 		$this->validateRequest( $request );
@@ -759,200 +709,6 @@ class UpdateServer {
 	 */
 	protected function normalizeFilePath( string $path ): string {
 		return \str_replace( [ \DIRECTORY_SEPARATOR, '\\' ], '/', $path );
-	}
-
-	/**
-	 * Write a log entry for the current request to the log file.
-	 *
-	 * @param Request $request The current API request.
-	 */
-	protected function logRequest( Request $request ): void {
-		$logFile = $this->getLogFileName();
-
-		$mustRotate = $this->logRotationEnabled && ! \file_exists( $logFile );
-
-		$handle = \fopen( $logFile, 'a' );
-		if ( $handle && \flock( $handle, \LOCK_EX ) ) {
-			$loggedIp = $request->clientIp;
-			if ( $this->ipAnonymizationEnabled ) {
-				$loggedIp = $this->anonymizeIp( $loggedIp );
-			}
-
-			$columns = [
-				'ip'                => $loggedIp,
-				'http_method'       => $request->httpMethod,
-				'action'            => $request->param( 'action', '-' ),
-				'slug'              => $request->param( 'slug', '-' ),
-				'installed_version' => $request->param( 'installed_version', '-' ),
-				'wp_version'        => $request->wpVersion ?? '-',
-				'site_url'          => $request->wpSiteUrl ?? '-',
-				'query'             => \http_build_query( $request->query, '', '&' ),
-			];
-			$columns = $this->filterLogInfo( $columns, $request );
-			$columns = $this->escapeLogInfo( $columns );
-
-			if ( isset( $columns['ip'] ) ) {
-				$columns['ip'] = \str_pad( $columns['ip'], 15, ' ' );
-			}
-			if ( isset( $columns['http_method'] ) ) {
-				$columns['http_method'] = \str_pad( $columns['http_method'], 4, ' ' );
-			}
-
-			$configuredTz = \ini_get( 'date.timezone' );
-			if ( empty( $configuredTz ) ) {
-				$defaultTz = \date_default_timezone_get();
-				if ( $defaultTz !== false ) {
-					\date_default_timezone_set( $defaultTz );
-				}
-			}
-
-			$line = \date( '[Y-m-d H:i:s O]' ) . ' ' . \implode( "\t", $columns ) . "\n";
-
-			\fwrite( $handle, $line );
-
-			if ( $mustRotate ) {
-				$this->rotateLogs();
-			}
-			\flock( $handle, \LOCK_UN );
-		}
-		if ( $handle ) {
-			\fclose( $handle );
-		}
-	}
-
-	/**
-	 * Build the full path to the current log file, including any rotation suffix.
-	 */
-	protected function getLogFileName(): string {
-		$path = $this->logDirectory . '/request';
-		if ( $this->logRotationEnabled ) {
-			$path .= '-' . \date( $this->logDateSuffix );
-		}
-		return $path . '.log';
-	}
-
-	/**
-	 * Adjust information that will be logged. Override in subclasses.
-	 *
-	 * @param array<string, string|null> $columns Key-value pairs of log data.
-	 * @param Request|null               $request The current API request, if available.
-	 * @return array<string, string|null> Filtered log columns.
-	 */
-	protected function filterLogInfo( array $columns, ?Request $request = null ): array {
-		return $columns;
-	}
-
-	/**
-	 * Escape all values in a log column array.
-	 *
-	 * @param string[] $columns Raw log column values.
-	 * @return string[] Escaped columns.
-	 */
-	protected function escapeLogInfo( array $columns ): array {
-		return \array_map( [ $this, 'escapeLogValue' ], $columns );
-	}
-
-	/**
-	 * Escape non-printable and non-graphic characters in a single log value.
-	 *
-	 * @param string|null $value The raw log value.
-	 */
-	protected function escapeLogValue( ?string $value ): ?string {
-		if ( $value === null ) {
-			return null;
-		}
-
-		$regex = '/[[:^graph:]]/';
-
-		if ( \function_exists( 'mb_check_encoding' ) && \mb_check_encoding( $value, 'UTF-8' ) ) {
-			$regex .= 'u';
-		}
-
-		$value = \str_replace( '\\', '\\\\', $value );
-		$value = \preg_replace_callback( $regex, [ $this, 'escapeNonGraphicCharacters' ], $value );
-
-		return $value;
-	}
-
-	/**
-	 * Convert non-graphic bytes in a regex match to hex escape sequences.
-	 *
-	 * @param array<int, string> $matches Regex match array.
-	 */
-	protected function escapeNonGraphicCharacters( array $matches ): string {
-		$length = \strlen( $matches[0] );
-		$escaped = '';
-		for ( $index = 0; $index < $length; $index++ ) {
-			$hexCode = \dechex( \ord( $matches[0][ $index ] ) );
-			$escaped .= '\x' . \strtoupper( \str_pad( $hexCode, 2, '0', \STR_PAD_LEFT ) );
-		}
-		return $escaped;
-	}
-
-	/**
-	 * Enable automatic log file rotation.
-	 *
-	 * @param string|null $rotationPeriod Date format for the log suffix; defaults to FILE_PER_MONTH.
-	 * @param int         $filesToKeep    Maximum number of rotated log files to retain.
-	 */
-	public function enableLogRotation( ?string $rotationPeriod = null, int $filesToKeep = 10 ): void {
-		if ( $rotationPeriod === null ) {
-			$rotationPeriod = self::FILE_PER_MONTH;
-		}
-
-		$this->logDateSuffix = $rotationPeriod;
-		$this->logBackupCount = $filesToKeep;
-		$this->logRotationEnabled = true;
-	}
-
-	/**
-	 * Delete old rotated log files that exceed the backup count.
-	 */
-	protected function rotateLogs(): void {
-		if ( $this->logBackupCount === 0 ) {
-			return;
-		}
-
-		$logFiles = \glob( $this->logDirectory . '/request*.log', \GLOB_NOESCAPE );
-		if ( \count( $logFiles ) <= $this->logBackupCount ) {
-			return;
-		}
-
-		\usort( $logFiles, 'strcmp' );
-		$logFiles = \array_reverse( $logFiles );
-
-		foreach ( \array_slice( $logFiles, $this->logBackupCount ) as $fileName ) {
-			if ( \is_file( $fileName ) && ! \unlink( $fileName ) ) {
-				\error_log( 'UpdateServer: Failed to delete rotated log file: ' . $fileName );
-			}
-		}
-	}
-
-	/**
-	 * Enable IP address anonymization in log entries.
-	 */
-	public function enableIpAnonymization(): void {
-		$this->ipAnonymizationEnabled = true;
-	}
-
-	/**
-	 * Mask the host portion of an IP address for privacy.
-	 *
-	 * @param string $ipAddress The IP address to anonymize.
-	 */
-	protected function anonymizeIp( string $ipAddress ): string {
-		$binaryIp = \inet_pton( $ipAddress );
-		if ( $binaryIp === false ) {
-			return $ipAddress;
-		}
-		if ( \strlen( $binaryIp ) === 4 ) {
-			$anonBinaryIp = $binaryIp & $this->ip4Mask;
-		} elseif ( \strlen( $binaryIp ) === 16 ) {
-			$anonBinaryIp = $binaryIp & $this->ip6Mask;
-		} else {
-			return $ipAddress;
-		}
-		return \inet_ntop( $anonBinaryIp );
 	}
 
 	/**
